requirements.txt
langchain>=0.3.0
langgraph>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
httpx>=0.28.0
icalendar>=6.1.0
python-dotenv>=1.0.1
fastmcp>=2.0.0
jinja2>=3.1.0
